requirements copy.txt
apiview==1.3.25
asgiref==3.5.1
autopep8==1.6.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-cors-headers==3.12.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.33.3
kiwisolver==1.4.2
MarkupPy==1.14
matplotlib==3.5.2
mysqlclient==2.1.0
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
paho-mqtt==1.6.1
Pillow==9.1.1
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
